+++ RequestNbp.py
@@ -1,5 +1,6 @@
 import requests
- 
+
+
 def get_historical_currency_rate(currency_code: str, currency_date: str) -> float:
     # date format is : 2000-12-01
  
@@ -7,5 +8,7 @@ def get_historical_currency_rate(currency_code: str, currency_date: str) -> floa
     r = url.json()
     rate = r['rates'][0]['mid']
     return rate
+
  
 print(get_historical_currency_rate('usd', '2012-05-18'))
+
